Fix new tile entities being removed 1 tick after set
Pasting a block with a tickable TileEntity with WorldEdit first sets the BlockState and then the TileEntity. Setting the BlockState already generates a default TileEntity. This is set in the world (loadedTileEntityList and tickableTileEntities), but also in the Chunk TileEntity map. When WorldEdit then sets the TileEntity to the proper entity, the first TileEntity is in some way invalidated, but not removed from the world. Minecraft later ticks the TileEntity through updateEntities() and finds out that the first TileEntity is invalid. Minecraft decides to remove it from loadedTileEntityList and tickableTileEntities, but also remove it by position in the Chunk map. Sadly, the removal by position removes the new TileEntity, causing the block to no longer have a TileEntity at all.

This fix first removes any existing tile entities at the position where a new one will be set, working around this problem in Minecraft.